requirements.txt
cfscrape==2.1.1
python-dateutil==2.8.0
python-dotenv==1.0.0
pytz==2019.2
requests==2.22.0
